# apps/api/openapi.yaml
openapi: 3.0.0
paths:
  /api/v1/auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "200":
          description: ""
      tags:
        - auth
  /api/v1/auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshDto"
      responses:
        "200":
          description: ""
      tags:
        - auth
  /api/v1/auth/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshDto"
      responses:
        "204":
          description: ""
      tags:
        - auth
  /api/v1/auth/change-password:
    post:
      operationId: AuthController_changePassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordDto"
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      tags:
        - auth
  /api/v1/auth/me:
    get:
      operationId: AuthController_me
      parameters: []
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      tags:
        - auth
  /api/v1/users:
    get:
      operationId: UsersController_list
      parameters:
        - name: q
          required: false
          in: query
          description: Busca por nombre o DNI
          schema:
            type: string
        - name: role
          required: false
          in: query
          schema:
            type: string
            enum:
              - ADMIN
              - OPERADOR
        - name: active
          required: false
          in: query
          schema:
            type: boolean
        - name: limit
          required: false
          in: query
          schema:
            default: 50
            type: number
        - name: cursor
          required: false
          in: query
          description: Cursor opaco devuelto en nextCursor
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      tags:
        - users
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: ""
      security:
        - bearer: []
      tags:
        - users
  /api/v1/users/{id}:
    get:
      operationId: UsersController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      tags:
        - users
    patch:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      tags:
        - users
  /api/v1/users/{id}/deactivate:
    post:
      operationId: UsersController_deactivate
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      tags:
        - users
  /api/v1/users/{id}/activate:
    post:
      operationId: UsersController_activate
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      tags:
        - users
  /api/v1/users/{id}/reset-password:
    post:
      operationId: UsersController_resetPassword
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - bearer: []
      tags:
        - users
  /health/live:
    get:
      operationId: HealthController_live
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - health
  /health/ready:
    get:
      operationId: HealthController_ready
      parameters: []
      responses:
        "200":
          description: The Health Check is successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  info:
                    type: object
                    example:
                      database:
                        status: up
                    additionalProperties:
                      type: object
                      required:
                        - status
                      properties:
                        status:
                          type: string
                      additionalProperties: true
                    nullable: true
                  error:
                    type: object
                    example: {}
                    additionalProperties:
                      type: object
                      required:
                        - status
                      properties:
                        status:
                          type: string
                      additionalProperties: true
                    nullable: true
                  details:
                    type: object
                    example:
                      database:
                        status: up
                    additionalProperties:
                      type: object
                      required:
                        - status
                      properties:
                        status:
                          type: string
                      additionalProperties: true
        "503":
          description: The Health Check is not successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  info:
                    type: object
                    example:
                      database:
                        status: up
                    additionalProperties:
                      type: object
                      required:
                        - status
                      properties:
                        status:
                          type: string
                      additionalProperties: true
                    nullable: true
                  error:
                    type: object
                    example:
                      redis:
                        status: down
                        message: Could not connect
                    additionalProperties:
                      type: object
                      required:
                        - status
                      properties:
                        status:
                          type: string
                      additionalProperties: true
                    nullable: true
                  details:
                    type: object
                    example:
                      database:
                        status: up
                      redis:
                        status: down
                        message: Could not connect
                    additionalProperties:
                      type: object
                      required:
                        - status
                      properties:
                        status:
                          type: string
                      additionalProperties: true
      tags:
        - health
info:
  title: A2C Inventario API
  description: Control de herramientas y maquinaria por obra — Constructora A2C
  version: 0.1.0
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    LoginDto:
      type: object
      properties:
        dni:
          type: string
          example: "30456789"
          description: DNI peruano de 8 dígitos
        password:
          type: string
        deviceName:
          type: string
          example: Galaxy A54
      required:
        - dni
        - password
    RefreshDto:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    ChangePasswordDto:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          description: Mínimo 8 caracteres, con letras y números
      required:
        - currentPassword
        - newPassword
    CreateUserDto:
      type: object
      properties:
        dni:
          type: string
          example: "35112604"
          description: DNI peruano de 8 dígitos
        fullName:
          type: string
          example: Martín Ruiz
        role:
          type: string
          enum:
            - ADMIN
            - OPERADOR
        temporaryPassword:
          type: string
          description: Contraseña temporal; el usuario deberá cambiarla en su primer ingreso
      required:
        - dni
        - fullName
        - role
        - temporaryPassword
    UpdateUserDto:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - OPERADOR